refactor: extract code that builds TenantConfOpt from requests into a From<> impl
